archive/.history/requirements_20241209164647.txt
requests>=2.31.0
pandas>=2.1.0
streamlit>=1.29.0
openai>=1.3.0
boto3>=1.34.0
aiohttp>=3.9.0
urllib3>=2.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
google>=3.0.0
fake-useragent>=1.4.0
SQLAlchemy>=2.0.0
tenacity>=8.2.0
email-validator>=2.1.0
streamlit-option-menu>=0.3.0
streamlit-tags>=1.2.0
plotly>=5.18.0
psycopg2-binary>=2.9.0
google-search-results>=2.4.2
